Refactor VHF code to separate submodule, fix normalization
